Add test and data for steadystate
